Using proxy Pattern for lua functions
