+++ src/lua/mod.rs
@@ -5,3 +5,4 @@ pub mod parsing;
 pub mod payloads;
 pub mod scan;
 pub mod threads;
+pub mod runtime;
